Store screen fixes
