+++ apps/dashboard/package.json
@@ -18,10 +18,10 @@
 		"@hookform/resolvers": "^5.0.1",
 		"@react-native-async-storage/async-storage": "2.2.0",
 		"@react-native-masked-view/masked-view": "0.3.2",
-		"@react-navigation/bottom-tabs": "^7.14.0",
-		"@react-navigation/elements": "^2.9.5",
-		"@react-navigation/native": "^7.1.28",
-		"@react-navigation/native-stack": "^7.13.0",
+		"@react-navigation/bottom-tabs": "^7.18.14",
+		"@react-navigation/elements": "^2.9.36",
+		"@react-navigation/native": "^7.3.14",
+		"@react-navigation/native-stack": "^7.18.6",
 		"@sentry/react-native": "~7.11.0",
 		"@shopify/flash-list": "2.0.2",
 		"@tanstack/react-query": "^5.75.0",
